requirements.txt
Flask==2.3.3
ipfshttpclient==0.8.0a2
requests==2.31.0
python-dotenv==1.0.0